webapp/requirements.txt
Flask==2.2.2
grpcio==1.42.0
matplotlib==3.8.0
networkx==3.2.1
protobuf==3.20.3
PyYAML==6.0.1
Werkzeug==2.2.3